gadget: start separating rule/convention validation from basic soundness
later this will be moved to a Validate function (current Validate will
become ValidateContent) and be done automatically only for newly
installed gadgets and at installation; not for gadgets already on the
system
