+++ gadget/validate_test.go
@@ -39,6 +39,124 @@ func (s *validateGadgetTestSuite) SetUpTest(c *C) {
 	s.dir = c.MkDir()
 }
 
+func (s *validateGadgetTestSuite) TestRuleValidateStructureReservedLabels(c *C) {
+	for _, tc := range []struct {
+		role, label, err string
+	}{
+		{label: "ubuntu-seed", err: `label "ubuntu-seed" is reserved`},
+		// 2020-12-02: disable for customer hotfix
+		/*{label: "ubuntu-boot", err: `label "ubuntu-boot" is reserved`},*/
+		{label: "ubuntu-data", err: `label "ubuntu-data" is reserved`},
+		{label: "ubuntu-save", err: `label "ubuntu-save" is reserved`},
+		// these are ok
+		{role: "system-boot", label: "ubuntu-boot"},
+		{label: "random-ubuntu-label"},
+	} {
+		err := gadget.RuleValidateVolumeStructure(&gadget.VolumeStructure{
+			Type:       "21686148-6449-6E6F-744E-656564454649",
+			Role:       tc.role,
+			Filesystem: "ext4",
+			Label:      tc.label,
+			Size:       10 * 1024,
+		})
+		if tc.err == "" {
+			c.Check(err, IsNil)
+		} else {
+			c.Check(err, ErrorMatches, tc.err)
+		}
+	}
+
+}
+
+func (s *validateGadgetTestSuite) TestEnsureVolumeRuleConsistency(c *C) {
+	state := func(seed bool, label string) *gadget.ValidationState {
+		systemDataVolume := &gadget.VolumeStructure{Label: label}
+		systemSeedVolume := (*gadget.VolumeStructure)(nil)
+		if seed {
+			systemSeedVolume = &gadget.VolumeStructure{}
+		}
+		return &gadget.ValidationState{
+			SystemSeed: systemSeedVolume,
+			SystemData: systemDataVolume,
+		}
+	}
+
+	for i, tc := range []struct {
+		s   *gadget.ValidationState
+		err string
+	}{
+
+		// we have the system-seed role
+		{state(true, ""), ""},
+		{state(true, "foobar"), "system-data structure must not have a label"},
+		{state(true, "writable"), "system-data structure must not have a label"},
+		{state(true, "ubuntu-data"), "system-data structure must not have a label"},
+
+		// we don't have the system-seed role (old systems)
+		{state(false, ""), ""}, // implicit is ok
+		{state(false, "foobar"), `.* must have an implicit label or "writable", not "foobar"`},
+		{state(false, "writable"), ""},
+		{state(false, "ubuntu-data"), `.* must have an implicit label or "writable", not "ubuntu-data"`},
+	} {
+		c.Logf("tc: %v %p %v", i, tc.s.SystemSeed, tc.s.SystemData.Label)
+
+		err := gadget.EnsureVolumeRuleConsistency(tc.s, nil)
+		if tc.err != "" {
+			c.Assert(err, ErrorMatches, tc.err)
+		} else {
+			c.Check(err, IsNil)
+		}
+	}
+
+	// Check system-seed label
+	for i, tc := range []struct {
+		l   string
+		err string
+	}{
+		{"", ""},
+		{"foobar", "system-seed structure must not have a label"},
+		{"ubuntu-seed", "system-seed structure must not have a label"},
+	} {
+		c.Logf("tc: %v %v", i, tc.l)
+		s := state(true, "")
+		s.SystemSeed.Label = tc.l
+		err := gadget.EnsureVolumeRuleConsistency(s, nil)
+		if tc.err != "" {
+			c.Assert(err, ErrorMatches, tc.err)
+		} else {
+			c.Check(err, IsNil)
+		}
+	}
+
+	// Check system-seed without system-data
+	vs := &gadget.ValidationState{}
+	err := gadget.EnsureVolumeRuleConsistency(vs, nil)
+	c.Assert(err, IsNil)
+	vs.SystemSeed = &gadget.VolumeStructure{}
+	err = gadget.EnsureVolumeRuleConsistency(vs, nil)
+	c.Assert(err, ErrorMatches, "the system-seed role requires system-data to be defined")
+
+	// Check system-save
+	vsWithSave := &gadget.ValidationState{
+		SystemData: &gadget.VolumeStructure{},
+		SystemSeed: &gadget.VolumeStructure{},
+		SystemSave: &gadget.VolumeStructure{},
+	}
+	err = gadget.EnsureVolumeRuleConsistency(vsWithSave, nil)
+	c.Assert(err, IsNil)
+	// use illegal label on system-save
+	vsWithSave.SystemSave.Label = "foo"
+	err = gadget.EnsureVolumeRuleConsistency(vsWithSave, nil)
+	c.Assert(err, ErrorMatches, "system-save structure must not have a label")
+	// complains when either system-seed or system-data is missing
+	vsWithSave.SystemSeed = nil
+	err = gadget.EnsureVolumeRuleConsistency(vsWithSave, nil)
+	c.Assert(err, ErrorMatches, "system-save requires system-seed and system-data structures")
+	vsWithSave.SystemData = nil
+	err = gadget.EnsureVolumeRuleConsistency(vsWithSave, nil)
+	c.Assert(err, ErrorMatches, "system-save requires system-seed and system-data structures")
+}
+
 func (s *validateGadgetTestSuite) TestValidateMissingRawContent(c *C) {
 	var gadgetYamlContent = `
 volumes:
